+++ spring-market/src/main/java/gb/spring/market/springmarket/contrllers/MainController.java
@@ -1,5 +1,6 @@
 package gb.spring.market.springmarket.contrllers;
 
+import gb.spring.market.springmarket.model.Product;
 import gb.spring.market.springmarket.services.CartService;
 import gb.spring.market.springmarket.services.ProductService;
 import gb.spring.market.springmarket.utils.UtilSessionFactory;
@@ -7,6 +8,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @Controller
 public class MainController {
@@ -20,30 +23,31 @@ public class MainController {
         this.cartService = cartService;
     }
 
-    @GetMapping("/new_product")
-    public String newProduct(Model model){
-        return "create_product";
-    }
 
-    @PostMapping("/new_product")
-    public String saveNewProduct(@RequestParam String title, @RequestParam int cost){
-        productService.createNewProduct(title, cost);
-        return "redirect:/";
-    }
-
-    @GetMapping("/info/{id}")
-    public String productInfoPage(Model model, @PathVariable long id){
-        model.addAttribute("product", productService.getProductByID(id));
-        return "productinfo";
-    }
-
-    @GetMapping("/delete/{id}")
-    public String deleteProduct(Model model, @PathVariable long id){
-        model.addAttribute("product", productService.getProductByID(id));
-        productService.deleteById(id);
-        return "delete";
-    }
-
+//    @GetMapping("/new_product")
+//    public String newProduct(Model model){
+//        return "create_product";
+//    }
+//
+//    @PostMapping("/new_product")
+//    public String saveNewProduct(@RequestParam String title, @RequestParam int cost){
+//        productService.createNewProduct(title, cost);
+//        return "redirect:/";
+//    }
+//
+//    @GetMapping("/info/{id}")
+//    public String productInfoPage(Model model, @PathVariable long id){
+//        model.addAttribute("product", productService.getProductByID(id));
+//        return "productinfo";
+//    }
+//
+//    @GetMapping("/delete/{id}")
+//    public String deleteProduct(Model model, @PathVariable long id){
+//        model.addAttribute("product", productService.getProductByID(id));
+//        productService.deleteById(id);
+//        return "delete";
+//    }
+//
     @GetMapping("/")
     public String main(Model model){
         model.addAttribute("items", productService.getProductList());
