+++ spring-market/src/main/resources/application.yaml
@@ -17,8 +17,8 @@ spring:
       hibernate:
                 dialect: org.hibernate.dialect.PostgreSQL94Dialect
                 #        dialect: org.hibernate.dialect.H2Dialect
-  h2:
-    console:
-      enabled: true
-      settings:
-        web-allow-others: false
+#  h2:
+#    console:
+#      enabled: true
+#      settings:
+#        web-allow-others: false
