+++ spring-market/src/main/resources/templates/index.html
@@ -41,4 +41,4 @@
     <a th:href="@{/new_product}">Create new product</a>
 </div>
 </body>
-</html>0
+</html>
